+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.4.6] WIP
+### [1.4.6] 2022-12-23
 
 -   Support OpenLayers 7.2.0
 -   Add `RLayerImage` for displaying a static image as a map